gitparser: Add exception handling of void function
Add exception handling code to determine whether there is actually
a problem with the function of void return type.
In the ground rule, void return type function should not be defined.
However, callback function is defined in other repository (including opensource project).
Therefore, we should follow format of callback function.

Upon examining the codes managed in our part, it is generally observed that for callback functions,
the suffixes "_cb", "_callback" or "_handler" are added to their names.
Therefore, these functions with such names will not be considered as violating the non-void function rule.

Signed-off-by: Unsung Lee <unsung.lee@samsung.com>